% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSEbootLS-package.R
\docType{package}
\name{LSEbootLS-package}
\alias{LSEbootLS}
\alias{LSEbootLS-package}
\title{Bootstrap Methods for Regression Models with Locally Stationary Errors}
\description{
Implements bootstrap methods for linear regression models with errors following a time-varying process, focusing on approximating the distribution of the least-squares estimator for regression models with locally stationary errors. It enables the construction of bootstrap and classical confidence intervals for regression coefficients, leveraging intensive simulation studies and real data analysis. The methodology is based on the approach described in Ferreira et al. (2020), allowing errors to be locally approximated by stationary processes.
}
\author{
\strong{Maintainer}: Nicolas Loyola \email{nloyola2016@udec.cl}

Authors:
\itemize{
  \item Guillermo Ferreira \email{gferreri@udec.cl}
  \item Joel Muñoz \email{joelmuno@udec.cl}
}

}
\keyword{internal}
